+++ packages/core/tarball/tests/convertDistRemoteToLocalTarballUrls.spec.ts
@@ -78,3 +78,41 @@ describe('convertDistRemoteToLocalTarballUrls', () => {
     );
   });
 });
+
+describe('convertDistRemoteToLocalTarballUrls - scoped', () => {
+  const fakeHost = 'fake.com';
+  const buildURI = (host, version) => `http://${host}/@org/npm_test/-/npm_test-${version}.tgz`;
+  const cloneMetadata = (pkg = metadata) => Object.assign({}, pkg);
+  const metadata: any = {
+    name: '@org/npm_test',
+    versions: {
+      '1.0.0': {
+        dist: {
+          tarball: 'http://registry.org/@org/npm_test/-/npm_test-1.0.0.tgz',
+        },
+      },
+      '1.0.1': {
+        dist: {
+          tarball: 'http://registry.org/@org/npm_test/-/npm_test-1.0.1.tgz',
+        },
+      },
+    },
+  };
+  test('should build a URI for dist tarball based on new domain', () => {
+    const req = httpMocks.createRequest({
+      method: 'GET',
+      headers: {
+        host: fakeHost,
+        [HEADERS.FORWARDED_PROTO]: 'http',
+      },
+      url: '/',
+    });
+    const convertDist = convertDistRemoteToLocalTarballUrls(cloneMetadata(), {
+      host: req.hostname,
+      headers: req.headers as any,
+      protocol: req.protocol,
+    });
+    expect(convertDist.versions['1.0.0'].dist.tarball).toEqual(buildURI(fakeHost, '1.0.0'));
+    expect(convertDist.versions['1.0.1'].dist.tarball).toEqual(buildURI(fakeHost, '1.0.1'));
+  });
+});