+++ packages/middleware/test/allow.spec.ts
@@ -1,22 +1,16 @@
 import request from 'supertest';
 
 import { HTTP_STATUS } from '@verdaccio/core';
-import { logger, setup } from '@verdaccio/logger';
 
 import { allow } from '../src';
 import { getApp } from './helper';
 
-setup({});
-
 test('should allow request', async () => {
-  const can = allow(
-    {
-      allow_publish: (params, remove, cb) => {
-        return cb(null, true);
-      },
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return cb(null, true);
     },
-    logger
-  );
+  });
   const app = getApp([]);
   // @ts-ignore
   app.get('/:package', can('publish'), (req, res) => {
@@ -27,14 +21,11 @@ test('should allow request', async () => {
 });
 
 test('should allow scope request', async () => {
-  const can = allow(
-    {
-      allow_publish: (params, remove, cb) => {
-        return cb(null, true);
-      },
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return cb(null, true);
     },
-    logger
-  );
+  });
   const app = getApp([]);
   // @ts-ignore
   app.get('/:package/:scope', can('publish'), (req, res) => {
@@ -45,14 +36,11 @@ test('should allow scope request', async () => {
 });
 
 test('should allow filename request', async () => {
-  const can = allow(
-    {
-      allow_publish: (params, remove, cb) => {
-        return cb(null, true);
-      },
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return cb(null, true);
     },
-    logger
-  );
+  });
   const app = getApp([]);
   // @ts-ignore
   app.get('/:filename', can('publish'), (req, res) => {
@@ -63,14 +51,11 @@ test('should allow filename request', async () => {
 });
 
 test('should not allow request', async () => {
-  const can = allow(
-    {
-      allow_publish: (params, remove, cb) => {
-        return cb(null, false);
-      },
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return cb(null, false);
     },
-    logger
-  );
+  });
   const app = getApp([]);
   // @ts-ignore
   app.get('/sec', can('publish'), (req, res) => {
@@ -81,17 +66,44 @@ test('should not allow request', async () => {
 });
 
 test('should handle error request', async () => {
-  const can = allow(
-    {
-      allow_publish: (params, remove, cb) => {
-        return cb(Error('foo error'));
-      },
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return cb(Error('foo error'));
     },
-    logger
-  );
+  });
   const app = getApp([]);
   // @ts-ignore
   app.get('/err', can('publish'));
 
   return request(app).get('/err').expect(HTTP_STATUS.INTERNAL_ERROR);
+});
+
+test('should allow request with version', async () => {
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return params.packageVersion === '1.0.0' ? cb(null, true) : cb(null, false);
+    },
+  });
+  const app = getApp([]);
+  // @ts-ignore
+  app.get('/:package/:version', can('publish'), (req, res) => {
+    res.status(HTTP_STATUS.OK).json({});
+  });
+
+  return request(app).get('/pacman/1.0.0').expect(HTTP_STATUS.OK);
+});
+
+test('should not allow request with version', async () => {
+  const can = allow({
+    allow_publish: (params, remove, cb) => {
+      return params.packageVersion === '1.0.0' ? cb(null, true) : cb(null, false);
+    },
+  });
+  const app = getApp([]);
+  // @ts-ignore
+  app.get('/:package/:version', can('publish'), (req, res) => {
+    res.status(HTTP_STATUS.OK).json({});
+  });
+
+  return request(app).get('/pacman/2.0.0').expect(HTTP_STATUS.INTERNAL_ERROR);
 });