+++ packages/loaders/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @verdaccio/loaders
 
+## 8.0.0-next-8.2
+
+### Minor Changes
+
+- 6a8154c: feat: update logger pino to latest
+
+### Patch Changes
+
+- Updated dependencies [6a8154c]
+  - @verdaccio/logger@8.0.0-next-8.2
+
 ## 8.0.0-next-8.1
 
 ### Patch Changes